Allow dropping buckets in the summary
This allows doing statistics on a subset of the buckets. It can be used to drop
overhead time (e.g. answering emails), to compare specific buckets (e.g.
ability to compare time spent doing a certain activity for different clients).

This is implemented by transferring the content of a bucket to the empty list.